Replace KDE system tray with ATP's own system tray
Replaces org.kde.plasma.systemtray with io.gitgud.wackyideas.systemtray.
The plasmoid now also needs to be compiled, and is also now the default
for ATP's session and layout templates.

This version of the system tray is also searchable in the gadgets window
for easier layout management. Everything else about the tray should be
largely the same as it was before.